Fix spotless not formatting all files
